refactor: move existing services into feature folders
